+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "karya-client"
-version = "0.1.0"
+version = "0.1.1"
 description = "Python client to interract with Karya; the open sourced distributed job scheduler"
 authors = ["Saumya Bhatt <108668549+SaumyaBhatt106@users.noreply.github.com>"]
 readme = "README.md"
