+++ samples/make_delayed_chained_karya_call.py
@@ -4,14 +4,14 @@ import os
 # Add the root directory of the project to the Python path
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from karya.commons.config import ClientConfig
-from karya.commons.entities.requests import (
+from karya.clients.config import ClientConfig
+from karya.clients import KaryaRestClient
+from karya.clients.requests import (
     CreateUserRequest,
     SubmitPlanRequest,
 )
-from karya.commons.entities.models.Action import RestApiRequest, ChainedRequest
-from karya.commons.entities.models.Plan import Recurring, OneTime
-from karya.clients.KaryaRestClient import KaryaRestClient
+from karya.entities.actions import RestApiRequest, ChainedRequest
+from karya.entities.plan_types import Recurring, OneTime
 
 
 async def main():
