+++ samples/make_recurring_push_to_kafka.py
@@ -4,14 +4,14 @@ import os
 # Add the root directory of the project to the Python path
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from karya.commons.config import ClientConfig
-from karya.commons.entities.requests import (
+from karya.clients.config import ClientConfig
+from karya.clients import KaryaRestClient
+from karya.clients.requests import (
     CreateUserRequest,
     SubmitPlanRequest,
 )
-from karya.commons.entities.models.Action import KafkaProducerRequest
-from karya.commons.entities.models.Plan import Recurring
-from karya.clients.KaryaRestClient import KaryaRestClient
+from karya.entities.actions import KafkaProducerRequest
+from karya.entities.plan_types import Recurring
 
 
 async def main():
